requirements.txt
tqdm==4.19.4
roman==3.0
spacy==2.0.7
python_Levenshtein==0.12.0
link==1.2.19
pytest==4.6.3
